% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bnd two factor}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{x}

\item{b}{x}
}
\value{
x
}
\description{
Create a new factor from two existing factors, where the new factors levls
are the union of the levels of th input factors.
}
\examples{
fbind(iris$Species[c(1, 51, 101)], PlantGrowth$group[c(1, 11, 21)])
}
