% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_into.R
\name{insert_into}
\alias{insert_into}
\title{Une data.frames con posición}
\usage{
insert_into(x, y, where = 1)
}
\arguments{
\item{x}{data.frame}

\item{y}{data.frame}

\item{where}{numeric lugar en que `y` se une a `x`}
}
\value{
data.frame
}
\description{
Une data.frames con posición
}
\examples{

t1 <- data.frame(a = 1, b = 2, c = 3)
t2 <- data.frame(x = 8, y = 9)

insert_into(t1, t2, where = 2)
}
